First pass at consumer
This is the first pass at a marshaled consumer. It uses the Marshal
system to do partition selection and claiming and then allows you to
consume messages from those partitions.

This is definitely nowhere near useful yet and I don't like the design
after writing it a bit, so I think I'm going to rework.